chore(forms): refactor forms audit worker tests
