+++ test/audits/forms/handler.test.js
@@ -1,0 +1,256 @@
+/*
+ * Copyright 2024 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+/* eslint-env mocha */
+
+import { expect, use } from 'chai';
+import sinon from 'sinon';
+import sinonChai from 'sinon-chai';
+import nock from 'nock';
+import {
+  formsAuditRunner,
+  processOpportunityStep,
+  runAuditAndSendUrlsForScrapingStep,
+} from '../../../src/forms-opportunities/handler.js';
+import { MockContextBuilder } from '../../shared.js';
+import formVitalsData from '../../fixtures/forms/formvitalsdata.json' with { type: 'json' };
+import expectedFormVitalsData from '../../fixtures/forms/expectedformvitalsdata.json' with { type: 'json' };
+import expectedFormSendToScraperData from '../../fixtures/forms/expectedformsendtoscraperdata.json' with { type: 'json' };
+import { FORM_OPPORTUNITY_TYPES } from '../../../src/forms-opportunities/constants.js';
+
+use(sinonChai);
+
+const sandbox = sinon.createSandbox();
+
+const baseURL = 'https://example.com';
+
+describe('Forms Vitals audit', () => {
+  const site = { getBaseURL: () => baseURL };
+
+  const context = new MockContextBuilder()
+    .withSandbox(sandbox)
+    .withOverrides({
+      runtime: { name: 'aws-lambda', region: 'us-east-1' },
+      func: { package: 'spacecat-services', version: 'ci', name: 'test' },
+      rumApiClient: {
+        queryMulti: sinon.stub().resolves(formVitalsData),
+      },
+    })
+    .build();
+
+  afterEach(() => {
+    nock.cleanAll();
+    sinon.restore();
+  });
+
+  it('form vitals audit runs rum api client formVitals query', async () => {
+    const FORMS_OPPTY_QUERIES = [
+      'cwv',
+      'form-vitals',
+    ];
+    const result = await formsAuditRunner(
+      'www.example.com',
+      context,
+      site,
+    );
+    expect(context.rumApiClient.queryMulti).calledWith(FORMS_OPPTY_QUERIES, {
+      domain: 'www.example.com',
+      interval: 15,
+      granularity: 'hourly',
+    });
+    expect(result).to.deep.equal(expectedFormVitalsData);
+  });
+});
+
+describe('audit and send scraping step', () => {
+  let context;
+  const siteId = 'test-site-id';
+
+  // eslint-disable-next-line prefer-const
+  context = new MockContextBuilder()
+    .withSandbox(sandbox)
+    .withOverrides({
+      runtime: { name: 'aws-lambda', region: 'us-east-1' },
+      func: { package: 'spacecat-services', version: 'ci', name: 'test' },
+      rumApiClient: {
+        queryMulti: sinon.stub().resolves(formVitalsData),
+      },
+      site: {
+        getId: sinon.stub().returns(siteId),
+      },
+      dataAccessStub: {
+        Opportunity: {
+          allBySiteIdAndStatus: sinon.stub().resolves([]),
+          create: sinon.stub(),
+        },
+      },
+      auditUrl: 'https://example.com',
+      formsOppty: {
+        getId: () => 'opportunity-id',
+        setAuditId: sinon.stub(),
+        save: sinon.stub(),
+        getType: () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
+      },
+      finalUrl: 'www.example.com',
+    })
+    .build();
+
+  afterEach(() => {
+    nock.cleanAll();
+    sinon.restore();
+  });
+
+  it('run audit and send urls for scraping step', async () => {
+    const FORMS_OPPTY_QUERIES = [
+      'cwv',
+      'form-vitals',
+    ];
+    const result = await runAuditAndSendUrlsForScrapingStep(context);
+    expect(context.rumApiClient.queryMulti).calledWith(FORMS_OPPTY_QUERIES, {
+      domain: 'www.example.com',
+      interval: 15,
+      granularity: 'hourly',
+    });
+    expect(result).to.deep.equal(expectedFormSendToScraperData);
+  });
+});
+
+describe('process opportunity step', () => {
+  let context;
+  const siteId = 'test-site-id';
+
+  let formsOppty = {
+    getId: () => 'opportunity-id',
+    setAuditId: sinon.stub(),
+    save: sinon.stub(),
+    getType: () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
+  };
+
+  let dataAccessStub = {
+    Opportunity: {
+      allBySiteIdAndStatus: sinon.stub().resolves([]),
+      create: sinon.stub().returns(formsOppty),
+    },
+  };
+
+  beforeEach(() => {
+    formsOppty = {
+      getId: () => 'opportunity-id',
+      setAuditId: sinon.stub(),
+      save: sinon.stub(),
+      getType: () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
+    };
+
+    dataAccessStub = {
+      Opportunity: {
+        allBySiteIdAndStatus: sinon.stub().resolves([]),
+        create: sinon.stub().returns(formsOppty),
+      },
+    };
+  });
+
+  // eslint-disable-next-line prefer-const
+  context = new MockContextBuilder()
+    .withSandbox(sandbox)
+    .withOverrides({
+      runtime: { name: 'aws-lambda', region: 'us-east-1' },
+      func: { package: 'spacecat-services', version: 'ci', name: 'test' },
+      rumApiClient: {
+        queryMulti: sinon.stub().resolves(formVitalsData),
+      },
+      site: {
+        getId: sinon.stub().returns(siteId),
+        getBaseURL: sinon.stub().returns('https://example.com'),
+        getLatestAuditByAuditType: sinon.stub().resolves({
+          auditResult: {
+            formVitals: [
+              {
+                url: 'https://example.com/form1',
+                pageview: { '2024-01-01': 1500 },
+                cwv: { lcp: 2.5, fid: 0.1 },
+              },
+            ],
+            auditContext: {
+              interval: 7,
+            },
+          },
+          fullAuditRef: 'www.example.com',
+          siteId: 'test-site-id',
+        }),
+        auditUrl: 'https://example.com',
+        formsOppty: {
+          getId: 'opportunity-id',
+          setAuditId: sinon.stub(),
+          save: sinon.stub(),
+          getType: FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
+        },
+      },
+      env: {
+        S3_SCRAPER_BUCKET_NAME: 'test-bucket',
+      },
+      dataAccess: dataAccessStub,
+      s3Client: {
+        send: sandbox.stub(),
+      },
+      finalUrl: 'www.example.com',
+    })
+    .build();
+
+  afterEach(() => {
+    nock.cleanAll();
+    sinon.restore();
+  });
+
+  it('process opportunity step', async () => {
+    context.s3Client.send.onCall(0).resolves({
+      Contents: [
+        { Key: 'scrapes/site-id/scrape.json' },
+      ],
+      IsTruncated: true,
+      NextContinuationToken: 'token',
+    });
+
+    context.s3Client.send.onCall(1).resolves({
+      Contents: [
+        { Key: 'scrapes/site-id/screenshot.png' },
+      ],
+      IsTruncated: false,
+      NextContinuationToken: 'token',
+    });
+
+    const mockFileResponse = {
+      ContentType: 'application/json',
+      Body: {
+        transformToString: sandbox.stub().resolves(JSON.stringify({
+          finalUrl: 'https://example.com/page1',
+          scrapeResult: {
+            rawBody: '<html lang="en"><body><header><a href="/home">Home</a><a href="/about">About</a></header></body></html>',
+            tags: {
+              title: 'Page 1 Title',
+              description: 'Page 1 Description',
+              h1: ['Page 1 H1'],
+            },
+          },
+        })),
+      },
+    };
+
+    context.s3Client.send.resolves(mockFileResponse);
+    formsOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION;
+    dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
+
+    const result = await processOpportunityStep(context);
+    expect(result).to.deep.equal({
+      status: 'complete',
+    });
+  });
+});
