+++ test/audits/forms/utils.test.js
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+/* eslint-env mocha */
+import { expect } from 'chai';
+import { shouldExcludeForm } from '../../../src/forms-opportunities/utils.js';
+
+describe('isSearchForm', () => {
+  it('should return true for search form type', () => {
+    const scrapedFormData = { formType: 'search' };
+    expect(shouldExcludeForm(scrapedFormData)).to.be.true;
+  });
+
+  it('should return true for login form type', () => {
+    const scrapedFormData = { formType: 'login' };
+    expect(shouldExcludeForm(scrapedFormData)).to.be.true;
+  });
+
+  it('should return true for form with unsubscribe class', () => {
+    const scrapedFormData = { classList: ['unsubscribe'] };
+    expect(shouldExcludeForm(scrapedFormData)).to.be.true;
+  });
+
+  it('should return false for non-search form', () => {
+    const scrapedFormData = {
+      formType: 'contact', classList: ['subscribe'], action: 'https://example.com/contact.html', fieldsLabels: ['Name', 'Email'],
+    };
+    expect(shouldExcludeForm(scrapedFormData)).to.be.false;
+  });
+});
